+++ src/main/java/com/michelin/cio/jenkins/plugin/requests/action/RequestDeleteFolderAction.java
@@ -39,6 +39,7 @@ import org.kohsuke.stapler.HttpRedirect;
 import org.kohsuke.stapler.HttpResponse;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
+import org.kohsuke.stapler.verb.POST;
 
 import com.michelin.cio.jenkins.plugin.requests.RequestsPlugin;
 import com.michelin.cio.jenkins.plugin.requests.model.DeleteFolderRequest;
@@ -75,15 +76,17 @@ public class RequestDeleteFolderAction extends FolderProperty<Folder> implements
 		errors.add(errorString);
 	}
 
+	@POST
 	public HttpResponse doCreateDeleteFolderRequest(StaplerRequest request,
 			StaplerResponse response)
 			throws IOException, ServletException, MessagingException {
+
 		try {
 			if (isIconDisplayed()) {
 				LOGGER.log(FINE, "Delete Folder Request");
 				errors.clear();
 				final String username = request.getParameter("username");
-				RequestsPlugin plugin = Jenkins.getInstance().getPlugin(RequestsPlugin.class);
+				RequestsPlugin plugin = Jenkins.get().getPlugin(RequestsPlugin.class);
 				String projectName = project.getFullName();
 				String projectFullName = project.getFullName();