+++ src/main/java/com/michelin/cio/jenkins/plugin/requests/model/RequestsUtility.java
@@ -42,7 +42,10 @@ import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
 import java.net.URI;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.logging.Logger;
 
 
@@ -138,5 +141,14 @@ public class RequestsUtility {
 		
 		return returnNameList;
 	}
+	
+	// Encode project name for url using `UTF-8` encoding scheme:
+    public String encodeValue (String value) {
+        try {
+            return URLEncoder.encode(value, StandardCharsets.UTF_8.toString());
+        } catch (UnsupportedEncodingException ex) {
+            throw new RuntimeException(ex.getCause());
+        }
+    }
 
 }